backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
pyjwt==2.9.0
bcrypt==4.3.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
